#294: Initial Carve out
