+++ lib/Builder.js
@@ -4,6 +4,7 @@ const TYPE = require('../types/mcdev.d');
 const Util = require('./util/util');
 const File = require('./util/file');
 const Cli = require('./util/cli');
+const Config = require('./util/config');
 const auth = require('./util/auth');
 const MetadataTypeInfo = require('./MetadataTypeInfo');
 // @ts-ignore
@@ -106,7 +107,7 @@ class Builder {
      */
     static async buildTemplate(businessUnit, selectedType, keyArr, market) {
         Util.logger.info('mcdev:: Build Definition from Template');
-        const properties = File.loadConfigFile();
+        const properties = await Config.getProperties();
         if (!Util._isValidType(selectedType)) {
             return;
         }
@@ -173,7 +174,7 @@ class Builder {
      * @returns {Promise.<void>} -
      */
     static async buildDefinition(businessUnit, selectedType, name, market) {
-        const properties = File.loadConfigFile();
+        const properties = await Config.getProperties();
         if (!Util._isValidType(selectedType)) {
             return;
         }
@@ -200,7 +201,7 @@ class Builder {
      * @returns {Promise.<void>} -
      */
     static async buildDefinitionBulk(listName, type, name) {
-        const properties = File.loadConfigFile();
+        const properties = await Config.getProperties();
         Util.verifyMarketList(listName, properties);
         if (type && !MetadataTypeInfo[type]) {
             Util.logger.error(`:: '${type}' is not a valid metadata type`);
