+++ lib/index.js
@@ -17,7 +17,7 @@ const MetadataTypeInfo = require('./MetadataTypeInfo');
 const MetadataTypeDefinitions = require('./MetadataTypeDefinitions');
 const Retriever = require('./Retriever');
 const cache = require('./util/cache');
-let properties;
+const Config = require('./util/config');
 
 /**
  * main class
@@ -43,10 +43,7 @@ class Mcdev {
      * @returns {void}
      */
     static async createDeltaPkg(argv) {
-        properties = properties || File.loadConfigFile();
-        if (!(await Util.checkProperties(properties))) {
-            return null;
-        }
+        const properties = await Config.getProperties();
         // get source market and source BU from config
         if (argv.filter) {
             return DevOps.getDeltaList(properties, argv.range, true, argv.filter);
@@ -60,10 +57,7 @@ class Mcdev {
      * @returns {Promise} .
      */
     static async selectTypes() {
-        properties = properties || File.loadConfigFile();
-        if (!(await Util.checkProperties(properties))) {
-            return null;
-        }
+        const properties = await Config.getProperties();
         await Cli.selectTypes(properties);
     }
     /**
@@ -77,7 +71,7 @@ class Mcdev {
      * @returns {Promise} .
      */
     static async upgrade(skipInteraction) {
-        properties = properties || File.loadConfigFile();
+        const properties = await Config.getProperties();
         if (!properties) {
             Util.logger.error('No config found. Please run mcdev init');
             return;
@@ -99,11 +93,7 @@ class Mcdev {
      */
     static async retrieve(businessUnit, selectedTypesArr, changelogOnly) {
         Util.logger.info('mcdev:: Retrieve');
-        properties = properties || File.loadConfigFile();
-        if (!(await Util.checkProperties(properties))) {
-            // return null here to avoid seeing 2 error messages for the same issue
-            return null;
-        }
+        const properties = await Config.getProperties();
 
         // assume a list was passed in and check each entry's validity
         if (selectedTypesArr) {
@@ -186,7 +176,7 @@ class Mcdev {
      * @returns {Promise.<object>} ensure that BUs are worked on sequentially
      */
     static async _retrieveBU(cred, bu, selectedTypesArr, changelogOnly) {
-        properties = properties || File.loadConfigFile();
+        const properties = await Config.getProperties();
         const buObject = await Cli.getCredentialObject(
             properties,
             cred !== null ? cred + '/' + bu : null,
@@ -273,7 +263,7 @@ class Mcdev {
     static async _deployBU(cred, bu, typeArr) {
         const buPath = `${cred}/${bu}`;
         Util.logger.info(`::Deploying ${buPath}`);
-        properties = properties || File.loadConfigFile();
+        const properties = await Config.getProperties();
         const buObject = await Cli.getCredentialObject(properties, buPath, null, true);
         if (buObject !== null) {
             cache.initCache(buObject);
@@ -297,7 +287,7 @@ class Mcdev {
      */
     static async deploy(businessUnit, selectedTypesArr) {
         Util.logger.info('mcdev:: Deploy');
-        properties = properties || File.loadConfigFile();
+        const properties = await Config.getProperties();
 
         if (selectedTypesArr) {
             for (const selectedType of selectedTypesArr) {
@@ -382,7 +372,7 @@ class Mcdev {
      */
     static async initProject(credentialsName, skipInteraction) {
         Util.logger.info('mcdev:: Setting up project');
-        properties = properties || File.loadConfigFile(!!credentialsName);
+        const properties = await Config.getProperties(!!credentialsName);
         await Init.initProject(properties, credentialsName, skipInteraction);
     }
 
@@ -394,7 +384,7 @@ class Mcdev {
      */
     static async findBUs(credentialsName) {
         Util.logger.info('mcdev:: Load BUs');
-        properties = properties || File.loadConfigFile();
+        const properties = await Config.getProperties();
         const buObject = await Cli.getCredentialObject(properties, credentialsName, true);
         if (buObject !== null) {
             BuHelper.refreshBUProperties(properties, buObject.credential);
@@ -410,7 +400,7 @@ class Mcdev {
      */
     static async document(businessUnit, type) {
         Util.logger.info('mcdev:: Document');
-        properties = properties || File.loadConfigFile();
+        const properties = await Config.getProperties();
         if (type && !MetadataTypeInfo[type]) {
             Util.logger.error(`:: '${type}' is not a valid metadata type`);
             return;
@@ -445,7 +435,7 @@ class Mcdev {
      */
     static async deleteByKey(businessUnit, type, customerKey) {
         Util.logger.info('mcdev:: delete');
-        properties = properties || File.loadConfigFile();
+        const properties = await Config.getProperties();
         const buObject = await Cli.getCredentialObject(properties, businessUnit);
         if (buObject !== null) {
             if ('string' !== typeof type) {
@@ -474,7 +464,7 @@ class Mcdev {
      * @returns {Promise.<void>} -
      */
     static async badKeys(businessUnit) {
-        properties = properties || File.loadConfigFile();
+        const properties = await Config.getProperties();
         const buObject = await Cli.getCredentialObject(properties, businessUnit);
         if (buObject !== null) {
             Util.logger.info('Gathering list of Name<>External Key mismatches (bad keys)');
@@ -553,7 +543,7 @@ class Mcdev {
      */
     static async retrieveAsTemplate(businessUnit, selectedType, name, market) {
         Util.logger.info('mcdev:: Retrieve as Template');
-        properties = properties || File.loadConfigFile();
+        const properties = await Config.getProperties();
         if (!Util._isValidType(selectedType)) {
             return;
         }
