+++ lib/util/cli.js
@@ -7,6 +7,7 @@ const inquirer = require('inquirer');
 const MetadataDefinitions = require('./../MetadataTypeDefinitions');
 const Util = require('./util');
 const auth = require('./auth');
+const Config = require('./config');
 
 /**
  * CLI helper class
@@ -38,7 +39,7 @@ const Cli = {
      * @returns {Promise.<void>} -
      */
     async addExtraCredential(properties, skipInteraction) {
-        if (!(await Util.checkProperties(properties))) {
+        if (!(await Config.checkProperties(properties))) {
             // return null here to avoid seeing 2 error messages for the same issue
             return null;
         } else {
@@ -85,7 +86,7 @@ const Cli = {
      */
     async getCredentialObject(properties, target, isCredentialOnly, allowAll) {
         try {
-            if (!(await Util.checkProperties(properties))) {
+            if (!(await Config.checkProperties(properties))) {
                 // return null here to avoid seeing 2 error messages for the same issue
                 return null;
             }
