+++ lib/util/file.js
@@ -443,57 +443,6 @@ const File = {
         }
     },
     /**
-     * loads central properties from config file
-     *
-     * @param {boolean} [silent] omit throwing errors and print messages; assuming not silent if not set
-     * @returns {object} central properties object
-     */
-    loadConfigFile(silent) {
-        let properties;
-        if (fs.existsSync(Util.configFileName)) {
-            try {
-                properties = fs.readJsonSync(Util.configFileName);
-            } catch (ex) {
-                Util.logger.error(`${ex.code}: ${ex.message}`);
-                return;
-            }
-            if (fs.existsSync(Util.authFileName)) {
-                let auth;
-                try {
-                    auth = fs.readJsonSync(Util.authFileName);
-                } catch (ex) {
-                    Util.logger.error(`${ex.code}: ${ex.message}`);
-                    return;
-                }
-
-                if (!auth) {
-                    const err = `${Util.authFileName} is not set up correctly.`;
-                    Util.logger.error(err);
-                    throw new Error(err);
-                }
-                for (const cred in properties.credentials) {
-                    if (auth[cred]) {
-                        if (properties.credentials[cred].eid !== auth[cred].account_id && !silent) {
-                            const err = `'${cred}' found in ${Util.configFileName} and ${Util.authFileName} have a Enterprise ID mismatch. Please check.`;
-                            Util.logger.error(err);
-                            throw new Error(err);
-                        }
-                        // TODO add auth checks #294
-                    } else if (!silent) {
-                        Util.logger.warn(
-                            `'${cred}' found in ${Util.configFileName} but not in ${Util.authFileName}. Please run 'mcdev init' to provide the missing credential details.`
-                        );
-                    }
-                }
-            } else if (!silent) {
-                Util.logger.warn(
-                    `${Util.authFileName} not found. Please run 'mcdev init' to provide the missing credential details.`
-                );
-            }
-        }
-        return properties;
-    },
-    /**
      * helper that splits the config back into auth & config parts to save them separately
      *
      * @param {object} properties central properties object
