+++ lib/util/init.config.js
@@ -6,6 +6,7 @@ const Util = require('./util');
 const inquirer = require('inquirer');
 const path = require('path');
 const semver = require('semver');
+const Config = require('./config');
 
 /**
  * CLI helper class
@@ -28,8 +29,8 @@ const Init = {
 
         const upgradeMsgs = [`Upgrading existing ${Util.configFileName}:`];
 
-        const missingFields = await Util.checkProperties(properties, true);
-        const defaultProps = Util.getDefaultProperties();
+        const missingFields = await Config.checkProperties(properties, true);
+        const defaultProps = Config.getDefaultProperties();
         if (missingFields.length) {
             missingFields.forEach((fieldName) => {
                 switch (fieldName) {
