+++ lib/util/init.js
@@ -7,6 +7,7 @@ const InitNpm = require('./init.npm');
 const InitConfig = require('./init.config');
 const inquirer = require('inquirer');
 const Util = require('./util');
+const Config = require('./config');
 
 /**
  * CLI helper class
@@ -52,7 +53,7 @@ const Init = {
                     }
                 } while (error && !skipInteraction);
                 Util.logger.debug('reloading config');
-                properties = File.loadConfigFile(true);
+                properties = await Config.getProperties(true);
             } else if (missingCredentials.length) {
                 // forced update-credential mode - user likely cloned repo and is missing mcdev-auth.json
                 Util.logger.warn(
@@ -82,7 +83,7 @@ const Init = {
                         }
                     } while (error);
                     Util.logger.debug('reloading config');
-                    properties = File.loadConfigFile(true);
+                    properties = await Config.getProperties(true);
                 }
                 Util.logger.info('✔️  All credentials updated.');
                 // assume node dependencies are not installed
